requirements-dev.txt
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
